Split MGS handler into gimlet/sidecar/psc/common
Building `mgmt-gateway` now requires setting exactly one of the cargo
features "gimlet"/"sidecar"/"psc". The latter two MGS handlers implement
common logic that already existed for gimlets (discovery, state,
updates, reset), but return "not supported" for all other calls.
