+++ task/mgmt-gateway/src/mgs_sidecar.rs
@@ -1,0 +1,160 @@
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at https://mozilla.org/MPL/2.0/.
+
+use core::convert::Infallible;
+
+use crate::{mgs_common::MgsCommon, Log, MgsMessage, __RINGBUF};
+use gateway_messages::{
+    sp_impl::SocketAddrV6, sp_impl::SpHandler, BulkIgnitionState,
+    DiscoverResponse, IgnitionCommand, IgnitionState, ResponseError,
+    SpComponent, SpPort, SpState, UpdateChunk, UpdateStart,
+};
+use ringbuf::ringbuf_entry;
+use task_net_api::UdpMetadata;
+
+pub(crate) struct MgsHandler {
+    common: MgsCommon,
+}
+
+impl MgsHandler {
+    /// Instantiate an `MgsHandler` that claims static buffers and device
+    /// resources. Can only be called once; will panic if called multiple times!
+    pub(crate) fn claim_static_resources() -> Self {
+        Self {
+            common: MgsCommon::claim_static_resources(),
+        }
+    }
+
+    pub(crate) fn drive_usart(&mut self) {}
+
+    pub(crate) fn wants_to_send_packet_to_mgs(&mut self) -> bool {
+        false
+    }
+
+    pub(crate) fn packet_to_mgs(
+        &mut self,
+        _tx_buf: &mut [u8; gateway_messages::MAX_SERIALIZED_SIZE],
+    ) -> Option<UdpMetadata> {
+        None
+    }
+}
+
+impl SpHandler for MgsHandler {
+    fn discover(
+        &mut self,
+        _sender: SocketAddrV6,
+        port: SpPort,
+    ) -> Result<DiscoverResponse, ResponseError> {
+        self.common.discover(port)
+    }
+
+    fn ignition_state(
+        &mut self,
+        _sender: SocketAddrV6,
+        _port: SpPort,
+        target: u8,
+    ) -> Result<IgnitionState, ResponseError> {
+        ringbuf_entry!(Log::MgsMessage(MgsMessage::IgnitionState { target }));
+        Err(ResponseError::RequestUnsupportedForSp)
+    }
+
+    fn bulk_ignition_state(
+        &mut self,
+        _sender: SocketAddrV6,
+        _port: SpPort,
+    ) -> Result<BulkIgnitionState, ResponseError> {
+        ringbuf_entry!(Log::MgsMessage(MgsMessage::BulkIgnitionState));
+        Err(ResponseError::RequestUnsupportedForSp)
+    }
+
+    fn ignition_command(
+        &mut self,
+        _sender: SocketAddrV6,
+        _port: SpPort,
+        target: u8,
+        command: IgnitionCommand,
+    ) -> Result<(), ResponseError> {
+        ringbuf_entry!(Log::MgsMessage(MgsMessage::IgnitionCommand {
+            target,
+            command
+        }));
+        Err(ResponseError::RequestUnsupportedForSp)
+    }
+
+    fn sp_state(
+        &mut self,
+        _sender: SocketAddrV6,
+        _port: SpPort,
+    ) -> Result<SpState, ResponseError> {
+        self.common.sp_state()
+    }
+
+    fn update_start(
+        &mut self,
+        _sender: SocketAddrV6,
+        _port: SpPort,
+        update: UpdateStart,
+    ) -> Result<(), ResponseError> {
+        self.common.update_start(update)
+    }
+
+    fn update_chunk(
+        &mut self,
+        _sender: SocketAddrV6,
+        _port: SpPort,
+        chunk: UpdateChunk,
+        data: &[u8],
+    ) -> Result<(), ResponseError> {
+        self.common.update_chunk(chunk, data)
+    }
+
+    fn serial_console_attach(
+        &mut self,
+        _sender: SocketAddrV6,
+        _port: SpPort,
+        _component: SpComponent,
+    ) -> Result<(), ResponseError> {
+        ringbuf_entry!(Log::MgsMessage(MgsMessage::SerialConsoleAttach));
+        Err(ResponseError::RequestUnsupportedForSp)
+    }
+
+    fn serial_console_write(
+        &mut self,
+        _sender: SocketAddrV6,
+        _port: SpPort,
+        offset: u64,
+        data: &[u8],
+    ) -> Result<u64, ResponseError> {
+        ringbuf_entry!(Log::MgsMessage(MgsMessage::SerialConsoleWrite {
+            offset,
+            length: data.len() as u16
+        }));
+        Err(ResponseError::RequestUnsupportedForSp)
+    }
+
+    fn serial_console_detach(
+        &mut self,
+        _sender: SocketAddrV6,
+        _port: SpPort,
+    ) -> Result<(), ResponseError> {
+        ringbuf_entry!(Log::MgsMessage(MgsMessage::SerialConsoleDetach));
+        Err(ResponseError::RequestUnsupportedForSp)
+    }
+
+    fn reset_prepare(
+        &mut self,
+        _sender: SocketAddrV6,
+        _port: SpPort,
+    ) -> Result<(), ResponseError> {
+        self.common.reset_prepare()
+    }
+
+    fn reset_trigger(
+        &mut self,
+        _sender: SocketAddrV6,
+        _port: SpPort,
+    ) -> Result<Infallible, ResponseError> {
+        self.common.reset_trigger()
+    }
+}
